Simplifying DtFolderStrategyFinder
